Reduce visual importance of Title
I was finding that Titles stood out too much compared to the rest of the
text due to the use of both a bright white foreground colour, as well as
bold. To remedy this I changed the colour of Title to be identical to
the one used for the normal text’s foreground.